+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Fix bug in `filter` of `datasets.load_datasets`
 * Fix the number of returned outliers when setting `n_outliers` to `None` for
   `KMeansLabeler`
+* Fix the number of outliers when setting `n_outliers` to `float` for
+  `MinorityLabeler`
 
 ## [v1.0.7]